requirements.txt
paho-mqtt
requests
python-telegram-bot
asyncio
flask
flask-cors
